+++ x-pack/plugins/encrypted_saved_objects/server/saved_objects/index.ts
@@ -23,7 +23,13 @@ interface SetupSavedObjectsParams {
   getStartServices: StartServicesAccessor;
 }
 
-export type SavedObjectsSetup = (includedHiddenTypes?: string[]) => EncryptedSavedObjectsClient;
+export type ClientInstanciator = (
+  options?: EncryptedSavedObjectsClientOptions
+) => EncryptedSavedObjectsClient;
+
+export interface EncryptedSavedObjectsClientOptions {
+  includedHiddenTypes?: string[];
+}
 
 export interface EncryptedSavedObjectsClient {
   getDecryptedAsInternalUser: <T = unknown>(
@@ -38,7 +44,7 @@ export function setupSavedObjects({
   savedObjects,
   security,
   getStartServices,
-}: SetupSavedObjectsParams): SavedObjectsSetup {
+}: SetupSavedObjectsParams): ClientInstanciator {
   // Register custom saved object client that will encrypt, decrypt and strip saved object
   // attributes where appropriate for any saved object repository request. We choose max possible
   // priority for this wrapper to allow all other wrappers to set proper `namespace` for the Saved
@@ -56,11 +62,11 @@ export function setupSavedObjects({
       })
   );
 
-  return (includedHiddenTypes?: string[]) => {
+  return clientOpts => {
     const internalRepositoryAndTypeRegistryPromise = getStartServices().then(
       ([core]) =>
         [
-          core.savedObjects.createInternalRepository(includedHiddenTypes),
+          core.savedObjects.createInternalRepository(clientOpts?.includedHiddenTypes),
           core.savedObjects.getTypeRegistry(),
         ] as [ISavedObjectsRepository, ISavedObjectTypeRegistry]
     );
