+++ x-pack/test/alerting_api_integration/security_and_spaces/tests/alerting/enable.ts
@@ -98,7 +98,9 @@ export default function createEnableAlertTests({ getService }: FtrProviderContex
           objectRemover.add(space.id, createdAlert.id, 'alert');
 
           await supertest
-            .put(`${getUrlPrefix(space.id)}/api/saved_objects/alert/${createdAlert.id}`)
+            .put(
+              `${getUrlPrefix(space.id)}/api/alerts_fixture/saved_object/alert/${createdAlert.id}`
+            )
             .set('kbn-xsrf', 'foo')
             .send({
               attributes: {
