+++ x-pack/test/encrypted_saved_objects_api_integration/fixtures/api_consumer_plugin/server/hidden_saved_object_routes.ts
@@ -26,7 +26,9 @@ export function registerHiddenSORoutes(
       try {
         return response.ok({
           body: await encryptedSavedObjects
-            .getClient([request.params.type])
+            .getClient({
+              includedHiddenTypes: [request.params.type],
+            })
             .getDecryptedAsInternalUser(request.params.type, request.params.id, { namespace }),
         });
       } catch (err) {
